[ADD] 삭제 구현
